+++ proc/mvm4/fu.go
@@ -10,14 +10,14 @@ import (
 type fetchUnit struct {
 	pc                 int32
 	l1i                l1i
-	remainingCycles    float32
+	remainingCycles    int
 	complete           bool
 	processing         bool
-	cyclesMemoryAccess float32
+	cyclesMemoryAccess int
 	toReset            bool
 }
 
-func NewFetchUnit(l1iCacheLineSizeInBytes int32, cyclesMemoryAccess float32) *fetchUnit {
+func NewFetchUnit(l1iCacheLineSizeInBytes int32, cyclesMemoryAccess int) *fetchUnit {
 	return &fetchUnit{
 		l1i:                newL1I(l1iCacheLineSizeInBytes),
 		cyclesMemoryAccess: cyclesMemoryAccess,
@@ -31,7 +31,7 @@ func (fu *fetchUnit) reset(pc int32) {
 	fu.toReset = true
 }
 
-func (fu *fetchUnit) cycle(currentCycle float32, app risc.Application, ctx *risc.Context, outBus comp.Bus[int]) {
+func (fu *fetchUnit) cycle(currentCycle int, app risc.Application, ctx *risc.Context, outBus comp.Bus[int]) {
 	// TODO Explain better why
 	// In case of a reset, we need to delete the last element in the comp.Bus
 	if fu.toReset {
